Constrain volume ID to ushort

+++ UnshieldSharp/InstallShieldCabinet.cs
@@ -250,7 +250,7 @@ namespace UnshieldSharp
                 ulong bytesToWrite = 0;
                 int result;
 
-                if (reader.VolumeBytesLeft == 0 && !reader.OpenVolume(reader.Volume + 1))
+                if (reader.VolumeBytesLeft == 0 && !reader.OpenVolume((ushort)(reader.Volume + 1)))
                 {
                     Console.Error.WriteLine($"Failed to open volume {reader.Volume + 1} to read {bytesLeft} more bytes");
                     reader.Dispose();
